[FIX]point of sale: passing analytic account
The statement in line 983 had no effect because the value of account_analytic_id
was being cleared by the product_id_change statement.
So in the end the invoice line was created without any account attached
Now it checks for an existing value, and if it is not already filled in,
the value from  _prepare_analytic_account is taken.