[IMP] stock : Improve return mechanism.
When a location is marked as return location,
upon a return it will put it in the possible return locations,
when it is a child location in the warehouse.

The push rules are also applied on returned moves now, but
only when the push rule does not return the return to the original location

The lot in the return line is removed as it does not make sense
(better to change it on the pack operations)
and the views are cleaned

The returned picking is also automatically reserved.