[IMP] stock: setting a warehouse partner on a picking can put the from/to location to transit
The onchange for location zones in picking takes into account
partner locations on partner of the picking.  (e.g. Suppliers/IT Vendors instead of Suppliers)
This is only the case when the default location of the picking type is empty.

If a partner is assigned to a warehouse, its default supplier/customer location becomes
the Transit location of the company.