[IMP] stock: Extra check of qties between pack operations and their lots detail + entire_pack param
As every pack operation can have several lots now,
we do an extra check if the qties in these lots details
match the total quantity of the pack operation.

There was an entire_pack key put in the context for the quants_move we want
to avoid by passing it as parameter.