[IMP] stock: product form Forecast statbutton with incoming/outgoing qty and product template and product product should be the same and consistent
The button Inventory Adjustement is also changed to Update Qty On Hand
as an inventory adjustment would expect a total inventory.

The incoming and outgoing quantity are important to see immediately
if there are some products to be added or not.

The stock move view (Traceability statbutton) should not allow for creating or editing immediately.