[WIP] mail: gateway: CIDs
Finally add support of CIDs in incoming emails. Inline images are recognized
and added as attachments. Image links are updated from src=cid: to src=link
using the /web/image controller.

It has been decided to keep the image in atttachments instead of putting them
in base64, like proposed in PR #TOFIND. Indeed we prefer to store this data
into the attachments instead of directly putting it in the mail_message table.
Moreover this enable attachments in the chatter and record views.