[IMP] web: remove useless RPC for ir_actions_server
This commit prevents the server to return an ir_actions_server to the
webclient as the webclient can't do anything with it, and is forced to
perform another RPC (/web/action/run) to execute the action. Instead, we
directly execute the server action, and return its resulting action if there
is one.