[FIX] point_of_sale: correctly copy the base_amount sign of fixed tax
All the POS has to do is copy what accounting does. In the case of a
fixed tax we should do (from account.tax._compute_amount):

return math.copysign(self.amount, base_amount) * quantity

where copysign(x, y) returns x with the sign of y.

When base_amount is 0 it's positive, not negative. This also correctly
handles a base_amount of -0, although it's something that we shouldn't
encounter in the POS.

opw-672118