[FIX] website : ir_qweb rendering when installing website
Sometimes, `request` does not have `website` attribute..

For example, create a fresh database, login and install
website_blog as first module. During the importation of
website_blog data, a qweb template rendering is required
(on create of blog.post). This will failed because we
are in a JsonRequest transation (by clicking on the
install button), and website has just been installed,
so the request object has no `website` attribute. But
website module extends ir.qweb engine, and check
`request.website` which cause traceback.
The problem only occur if a qweb template needs to be
rendered during the same request installing website module.

This fix propose the same solution already in place in
`render` method of ir_ui_view.py of website module :
checking the existance of `website` attribute.