[FIX] account_voucher: round move line with company currency
When creating move line for voucher, only the total amount is rounded
to the voucher company currency. Thus in some instance the voucher move
lines would not be balanced.

This commit add a key in the context so move line may be rounded.

The view before validating will still not be rounded, but : there is a
currency on the voucher and a currency on the voucher's company (used
for rounding when creating move line) so it would anyway be inconsistent
if these two currencies are different.

opw-673255