[IMP] base, web: add type=fullscreen to ir_actions
With this commit, actions can have a type=fullscreen.  This means that
the navbar will be hidden, for the action and for all subsequent actions
open in the breadcrumb

There is no real security behind this, this is just a feature for
convenience, so users can bookmark an url an open some action in a way
that they are not distracted by anything.  In short, this is some kind of
light 'kiosk' mode.