[REV] Packaging: remove "openerp-gevent" from script
This reverts commit a0eb172cab590982168492ad16ad90c2228c6424 and thus
allow to actually use the longpolling feature.

A better solution will be implemented for v10: removing the `openerp-gevent`
script to replace it with a special argument in `odoo.py`.

There is still a dependency issue as we need psycogreen but this dependency
is not met in debian wheezy. This will be made a hard dependency for v10 as
it is available in debian stretch.

opw 678334