[FIX] point_of_sale: restored removed report
The sale of the day report was removed during new api migration as it was
thought, it could be replaced by pivot view.
Due to the lack of some options and the problems to be printed, create similar
report.

This commit adds a report in the PoS interface to be printed on ESC/POS printers