[FIX] web_tour: ideal location to append tip
The tip was currently placed in the first parent which can scroll.
This is wrong if there is parent which is position:fixed in that
element which scroll (as it would not follow the scrolling as it is
fixed).
The ideal location is now the first parent which can scroll or which
is position:fixed.

Note: there are still two cases which may not work but they are
considered rare enough to not treat them :
* A position:fixed element which also is overflow:hidden
* A position:absolute element whose relative ancestor is an ancestor
of the ancestor element which can scroll.