[FIX] *: html fields sanitization
Since commit c9e38cae9f88bb61fd4468ad12e3e3200a4498ac in saas-10, the html fields are sanitized in
strict mode by default. This breaks all html fields behaviour when it
is editable in the website editor.

This commit deactivate the sanitization for such HTML fields (like it
was already the case for blogs contents or products descriptions for
example).