[IMP] mail,web,web_planner: order of systray items
Sort systray items by sequence to allow, e.g. to force the planner item to
be the left-most item of the systray as it is not displayed in all apps, and
thus produces a flickering of the other items.

Also invert the push order of the two mail items to keep the previous order
as they are now prepended instead of appended.