[FIX] account: fix full reconciliation.
when doing a full reconciliation in the same currency, full_reconcile_id was not set on account_partial_reconcile object
resulting in problem when trying to unreconcile entries.

Also add matching number to the exchanged rate entries created in case of full reconciliation.

''this fix is specific to v9, another one has been submitted in saas-10''