[CLEAN] mail: mail gateway code
The purpose of this commit is to clean a bit the code related to mail routing.
Indeed it is now clearly separeted into two categories. First one is when
the incoming email is an answer to an existing thread. Second one is when
the incoming email is not an answer and will create a new thread.