[IMP] mail: clean test about forwarded emails and parent_id
Previous version of the test was actually not correct. Mailgateway was not
considering references as in-reply-to when trying to find answers that have
to be considered as a new thread and not an answer to an existing thread.

Purpose of the test is actually to be sure that answers considered as a new
thread have no parent_id set. Otherwise you may have threads linked to threads
in another record, leading to access right issues. Moreover it this message
begins a new thread it makes effectively no sense to have a parent_id.

New test is written to better simulate the scope. It sets a message_id
containing reply_to which is the way odoo uses to indicated that answers
to a message have to be considered as new threads.